[Lesson 5] Имплиментировали интерефейс OnWeatherLoaded в DetailsFragment, создали поток в loadWeather

+++ app/src/main/java/com/example/kotlin_lessons/model/WeatherDTO.kt
@@ -117,7 +117,7 @@ class WeatherDTO (
 
     data class Info (
         val url: String,
-        val lat: Long,
-        val lon: Long
+        val lat: Double,
+        val lon: Double
     )
